feat: 최종주차 프로젝트 진행 상황 적용

+++ src/components/mentoring/dashboard/ContactTabMenteeSide.jsx
@@ -1,29 +1,28 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { useQueryClient, useQuery, useMutation } from "@tanstack/react-query";
 import toast from "react-hot-toast";
+import { useAtomValue } from "jotai";
 
 import {
   getContactConnectionsReq,
   deleteConnectionReq,
 } from "../../../apis/mentoring/connetion";
 import { convertDateToAge } from "../../../utils/age";
+import { profileImageAtom } from "../../../store";
 
 import FlagTag from "../../common/FlagTag";
 import Tag from "../../common/Tag";
 import Button from "../../common/Button";
-import Fallback from "../../common/Fallback";
-import Loader from "../../common/Loader";
-import Error from "../../common/Error";
-import ProfileModal from "./ProfileModal";
+import CreateProfileModal from "./CreateProfileModal";
+import NotPost from "./NotPost";
 
 export default function ContactTabMenteeSide() {
   const navigate = useNavigate();
   const queryClient = useQueryClient();
-
   const [isModal, setIsModal] = useState(false);
-
   const [modalUid, setModalUid] = useState(null);
+  const defaultImage = useAtomValue(profileImageAtom);
 
   const { data } = useQuery({
     queryKey: ["contacts"],
@@ -55,27 +54,34 @@ export default function ContactTabMenteeSide() {
   };
 
   const handleCancelClick = () => {
-    if (window.confirm("Are you sure you want to cancel?")) {
-      deleteMutate(
-        Object.keys(checks).reduce((acc, key) => {
-          if (checks[key] === true) return [...acc, key];
-          else return [...acc];
-        }, []),
-        {
-          onSuccess: () => {
-            toast("Successfully canceled.");
-            queryClient.invalidateQueries({
-              queryKey: ["contacts"],
-            });
-            setChecks(
-              data.data.data.reduce((acc, post) => {
-                return { ...acc, [post.connectionId]: false };
-              }, {})
-            );
-          },
-        }
-      );
-    } else toast("No mentors have been selected.");
+    if (Object.values(checks).some((val) => val)) {
+      if (window.confirm("Are you sure you want to cancel?")) {
+        deleteMutate(
+          Object.keys(checks).reduce((acc, key) => {
+            if (checks[key] === true) return [...acc, key];
+            else return [...acc];
+          }, []),
+          {
+            onSuccess: () => {
+              toast("Successfully canceled.", {
+                className: "bg-[#5A906E] text-[#F2F7F5]",
+              });
+              queryClient.invalidateQueries({
+                queryKey: ["contacts"],
+              });
+              setChecks(
+                data.data.data.reduce((acc, post) => {
+                  return { ...acc, [post.connectionId]: false };
+                }, {})
+              );
+            },
+          }
+        );
+      }
+    } else
+      toast("No mentors have been selected.", {
+        className: "bg-[#5A906E] text-[#F2F7F5]",
+      });
   };
 
   const handlePostClick = (e) => {
@@ -100,7 +106,10 @@ export default function ContactTabMenteeSide() {
                 type="checkbox"
                 name="all"
                 className="accent-green-600"
-                checked={Object.values(checks).every((val) => val === true)}
+                checked={
+                  Object.values(checks).length !== 0 &&
+                  Object.values(checks).every((val) => val === true)
+                }
                 onChange={handleCheckBoxChange}
               />
             </th>
@@ -131,8 +140,8 @@ export default function ContactTabMenteeSide() {
               </td>
               <td className="p-2 text-left space-x-2 flex items-center">
                 <img
-                  className="inline w-8 rounded-full"
-                  src={post.writerDTO.profileImage}
+                  className="inline object-fill w-8 h-8 rounded-full"
+                  src={post.writerDTO.profileImage || defaultImage}
                   alt={`${post.writerDTO.mentorId} 프로필 이미지`}
                 ></img>
                 <div className="inline-flex flex-col">
@@ -173,18 +182,17 @@ export default function ContactTabMenteeSide() {
           ))}
         </tbody>
       </table>
+      {data.data.data.length === 0 && <NotPost />}
       <div className="mt-4 text-right">
         <Button color="white" size="sm" onClick={handleCancelClick}>
           Cancel
         </Button>
       </div>
-      <Fallback Loader={Loader} Error={Error} errorMessage="ERROR">
-        <ProfileModal
-          isModal={isModal}
-          setIsModal={setIsModal}
-          uid={modalUid}
-        />
-      </Fallback>
+      <CreateProfileModal
+        isModal={isModal}
+        setIsModal={setIsModal}
+        uid={modalUid}
+      />
     </>
   );
 }
